pyproject.toml
[tool.poetry]
name = "pyfy"
version = "0.1.0"
description = "Python Finance"
authors = ["Wayne Moorefield <wayne.moorefield@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.3"
flake8 = "^7.1.1"
flake8-bugbear = "^24.8.19"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
google-sheets-test = "pyfy.sheets:test"
